% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{CI}
\alias{CI}
\title{Confidence interval equivalence test}
\usage{
CI(model, mmd, term, alpha = 0.05)
}
\arguments{
\item{model}{lmer object}

\item{mmd}{minimal meaningful difference}

\item{term}{model term}

\item{alpha}{desired alpha}
}
\value{
list of results
}
\description{
Compute an equivalence test using a symmetric confidence interval approach
}
