% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{evalci}
\alias{evalci}
\title{Compare MMD against confidence interval}
\usage{
evalci(ci, mmd)
}
\arguments{
\item{ci}{Named numeric vector with "lower" and "upper" elements}

\item{mmd}{Minimal meaningful difference}
}
\value{
Logical value indicating whether MMD is within the CI
}
\description{
Compare MMD against confidence interval
}
