% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lmer_values}
\alias{get_lmer_values}
\title{Get values from lmer object}
\usage{
get_lmer_values(model, term, alpha = 0.05)
}
\arguments{
\item{model}{lmer object}

\item{term}{model term}

\item{alpha}{desired alpha value}
}
\value{
list of values
}
\description{
Extract a specific model parameter's estimate, degrees of freedom and error
}
