% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.uppercase}
\alias{is.uppercase}
\title{Detects if all characters in a character are uppercase}
\usage{
is.uppercase(x)
}
\arguments{
\item{x}{character}
}
\value{
boolean
}
\description{
Detects if all characters in a character are uppercase
}
