% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mixequivResults}
\alias{summary.mixequivResults}
\title{Summary for equivalence test result}
\usage{
\method{summary}{mixequivResults}(object, ...)
}
\arguments{
\item{object}{An object of class "mixequivResults"}

\item{...}{further arguments (ignored)}
}
\value{
A printed summary
}
\description{
Summary for equivalence test result
}
