% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tost.R
\name{TOST}
\alias{TOST}
\title{Two one-sided t-test}
\usage{
TOST(model, mmd, term, alpha = 0.05)
}
\arguments{
\item{model}{lmer object}

\item{mmd}{minimal meaningful difference}

\item{term}{model term}

\item{alpha}{desired alpha value}
}
\value{
list of results
}
\description{
Compute two one-sided t-test with a desired alpha and minimal meaningful
difference.
}
