% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tozt.R
\name{TOZT}
\alias{TOZT}
\title{Two one sided z-test (Wald test)}
\usage{
TOZT(model, mmd, term, alpha)
}
\arguments{
\item{model}{lmer object}

\item{mmd}{minimal meaningful difference}

\item{term}{model term}

\item{alpha}{desired alpha value}
}
\value{
list of results
}
\description{
Compute two one-sided z-test with a desired alpha and minimal meaningful
difference.
}
